Add a for loop to create a grid of squares
The first for loop creates a row with an assigned class name per row.
The first class name is to individualize the each row, the second class
is to apply styling for all rows. The second loop creates the squares
for each row and again names each individually and also applies a
class for consistent styling.